requirements.txt
Flask==3.0.3
Pillow==10.4.0
Click==8.1.7